requirements.txt
requests
psutil
pywin32
pillow
tkinter-tooltip
cx_Freeze
obsws-python
obs-websocket-py
pytesseract
opencv-python
ultralytics
transformers
torch
